Restructure views
Tidy up `app.py` to make it more readable
by exporting views into `views.py` and `login.py`.
