+++ webapp/macaroons.py
@@ -1,7 +1,28 @@
-from openid.extension import Extension as OpenIDExtension
+# Standard library
+import base64
+
+# Packages
+from openid.extension import Extension
 
 
-class MacaroonRequest(OpenIDExtension):
+def binary_serialize_macaroons(macaroons):
+    """Encode all serialized macaroons and concatonate as a serialize bytes
+    @param macaroons: Iterable of macaroons lead by root_macaroon as first
+       element followed by any discharge macaroons to serialize
+    """
+    serialized_macaroons = []
+    for macaroon in macaroons:
+        serialized = macaroon.serialize()
+        encoded = serialized.encode("utf-8")
+        padded = encoded + b"=" * (-len(encoded) % 4)
+        serialized_macaroons.append(base64.urlsafe_b64decode(padded))
+
+    serialized = base64.urlsafe_b64encode(b"".join(serialized_macaroons))
+
+    return serialized + b"=" * (-len(serialized) % 4)
+
+
+class MacaroonRequest(Extension):
     ns_uri = "http://ns.login.ubuntu.com/2016/openid-macaroon"
     ns_alias = "macaroon"
 
@@ -16,7 +37,7 @@ class MacaroonRequest(OpenIDExtension):
         return {"caveat_id": self.caveat_id}
 
 
-class MacaroonResponse(OpenIDExtension):
+class MacaroonResponse(Extension):
     ns_uri = "http://ns.login.ubuntu.com/2016/openid-macaroon"
     ns_alias = "macaroon"
